Testing infrastructure fixes
New, undocumented-on-purpose -mocktime=timestamp command-line
argument to startup with mocktime set. Needed because
time-related blockchain sanity checks are done on startup, before a
test has a chance to make a setmocktime RPC call.

And changed the setmocktime RPC call so calling it will not result in
currently connected peers being disconnected due to inactivity timeouts.

And fixed a -regest bug:

Blocks generated in -regtest mode would pay to a key that was left
in the keypool, which can cause subtle bugs. In particular, I expected
every block generated to be unique, but got duplicate blocks between
calls to setgenerate.

Fixes Bitcoin Core issue 6268.