Make head-first mining configurable via getblocktemplate
This makes head-first mining opt-in. To get empty
blocks built on validated headers, pass:

    { "mode" : "template", "headfirst" : 30 }

... in the getblocktemplate call. The argument
to "headfirst" is the maximum number of seconds
getblocktemplate will return an empty block between
receiving and validating the block header and receiving
and validating the full block data.

A new parameter is added to the getblocktemplate result
object:

    "previousfullyvalidated" : true|false

... which will be false if the template returned is an
empty block built on one or more headers.